+++ ash/src/vk/definitions.rs
@@ -2,7 +2,8 @@ use crate::vk::aliases::*;
 use crate::vk::bitflags::*;
 use crate::vk::constants::*;
 use crate::vk::enums::*;
-use crate::vk::prelude::*;
+use crate::vk::platform_types::*;
+use crate::vk::{ptr_chain_iter, Handle};
 use std::fmt;
 use std::os::raw::*;
 #[doc = "<https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/VkSampleMask.html>"]
